requirements.txt
anthropic
beautifulsoup4
boto3
celery
channels
channels-redis
daphne
Django
dj-rest-auth
django-allauth
django-cors-headers
django-countries
django-extensions
django-filter
django-ratelimit
django-rest-passwordreset
django-rest-swagger
django-storages
djangorestframework
djangorestframework-simplejwt
dj-database-url
drf-spectacular
drf-yasg
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
langchain
langchain-community
langchain-openai
llama-index
markdown
more-itertools
numpy
openai
opencv-python
openpyxl
pandas
Pillow
pyspellchecker
playwright
python-decouple
python-dotenv
python-Levenshtein
psycopg2-binary
scikit-learn
selenium
stripe
